requirements.txt
wxPython
selenium
requests
pandas
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pygetwindow
pywin32 